@@ -29,17 +29,19 @@
   "homepage": "https://github.com/cesarandreu/react-window-resize-listener#readme",
   "devDependencies": {
     "phantomjs": "^1.9.18",
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0",
     "standard": "^5.3.1",
     "tape": "^4.2.2",
     "zuul": "^3.7.1"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
   },
   "dependencies": {
-    "lodash.debounce": "^3.1.1"
+    "create-react-class": "^15.6.3",
+    "lodash.debounce": "^3.1.1",
+    "prop-types": "^15.7.2"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
+    "react": "^0.14.0 || ^15.0.0 || ^16.0.0"
   },
   "directories": {
     "test": "test"
