@@ -4,8 +4,10 @@
  */
 var React = require('react')
 var debounce = require('lodash.debounce')
+var PropTypes = require('prop-types')
+var createClass = require('create-react-class')
 
-var WindowResizeListener = React.createClass({
+var WindowResizeListener = createClass({
   displayName: 'WindowResizeListener',
 
   propTypes: {
@@ -14,7 +16,7 @@ var WindowResizeListener = React.createClass({
      * type WindowSize = { windowWidth: number, windowHeight: number }
      * type onResize = (windowSize: WindowSize) => void
      */
-    onResize: React.PropTypes.func.isRequired
+    onResize: PropTypes.func.isRequired
   },
 
   statics: {
